+++ app/src/main/java/com/archive/mynews/model/Article.kt
@@ -1,0 +1,14 @@
+package com.archive.mynews.model
+
+import java.util.*
+
+data class Article(
+    val source: Source,
+    val author: String,
+    val title: String,
+    val description: String,
+    val url: String,
+    val urlToImage: String,
+    val publishedAt: Date,
+    val content: String
+)
