+++ app/src/main/java/com/archive/mynews/model/CountryCode.kt
@@ -1,0 +1,5 @@
+package com.archive.mynews.model
+
+enum class CountryCode(val code: String) {
+    REPUBLIC_OF_KOREA("kr"),
+}
