+++ app/src/main/java/com/archive/mynews/model/Source.kt
@@ -1,0 +1,11 @@
+package com.archive.mynews.model
+
+data class Source(
+    val id: String,
+    val name: String,
+    val description: String,
+    val url: String,
+    val category: Category,
+    val language: String,
+    val country: String
+)
